medusa-cloud-components-python-backend/medusa-cloud-components-python-backend/backend-py/requirements.txt
fastapi==0.115.2
mangum==0.17.0
boto3==1.35.36
bcrypt==4.2.0
PyJWT==2.9.0
uvicorn==0.32.0
